requirements.txt
anyio==3.6.2
astroid==2.15.5
Babel==2.12.1
blinker==1.6.2
certifi==2022.12.7
click==8.1.3
dill==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.1
Flask==2.3.2
Flask-Admin==1.6.1
flask-babel==3.1.0
Flask-BabelEx==0.9.4
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-WTF==1.1.1
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
numpy==1.24.3
orjson==3.8.11
packaging==23.1
pandas==2.0.1
platformdirs==3.8.0
pluggy==1.2.0
pydantic==1.10.7
pylint==2.17.4
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
speaklater==1.3
sqladmin==0.10.3
SQLAlchemy==2.0.12
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
Werkzeug==2.3.3
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
